Add user authentication and ownership middleware functions
